pyproject.toml
[project]
name = "p4-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "lightrag-hku>=1.3.8",
    "neo4j>=5.28.1",
    "networkx>=3.5",
    "ollama>=0.5.1",
    "pymupdf>=1.26.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
